+++ src/client/App.tsx
@@ -12,8 +12,7 @@ import Home from './pages/Home';
 const Product = lazy(() => import('./pages/Product'));
 const About = lazy(() => import('./pages/About'));
 const Login = lazy(() => import('./pages/Login'));
-const Privacy = lazy(() => import('./pages/Privacy'));
-const Terms = lazy(() => import('./pages/Terms'));
+const Policies = lazy(() => import('./pages/Policies'));
 const NotFound = lazy(() => import('./pages/NotFound'));
 
 const App = () => {
@@ -36,8 +35,7 @@ const App = () => {
           <Route path={ROUTES.PRODUCT} element={<Product />} />
           <Route path={ROUTES.ABOUT} element={<About />} />
           <Route path={ROUTES.LOGIN} element={<Login />} />
-          <Route path={ROUTES.PRIVACY} element={<Privacy />} />
-          <Route path={ROUTES.TERMS} element={<Terms />} />
+          <Route path={ROUTES.POLICIES} element={<Policies />} />
           <Route path="*" element={<NotFound />} />
         </Routes>
       </BrowserRouter>